+++ packages/blade/src/components/Input/BaseInput/BaseInput.tsx
@@ -361,6 +361,10 @@ type BaseInputCommonProps = FormInputLabelProps &
      * @default undefined
      */
     tabIndex?: number;
+    /**
+     * Slot to be rendered adjacent to the value
+     */
+    valueSuffix?: React.ReactNode;
   } & TestID &
   Platform.Select<{
     native: {
@@ -866,6 +870,7 @@ const _BaseInput: React.ForwardRefRenderFunction<BladeElementRef, BaseInputProps
     trailingDropDown,
     labelSuffix,
     labelTrailing,
+    valueSuffix,
     ...rest
   },
   ref,
@@ -1093,6 +1098,7 @@ const _BaseInput: React.ForwardRefRenderFunction<BladeElementRef, BaseInputProps
                 trailingInteractionElement={trailingInteractionElement}
                 leadingInteractionElement={leadingInteractionElement}
                 suffix={suffix}
+                valueSuffix={valueSuffix}
                 trailingIcon={trailingIcon}
                 maxCharacters={maxCharacters}
                 textAlign={textAlign}